services/rag-server/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain & AI
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0

# Vector Store
chromadb>=0.4.0

# Document Processing
pypdf>=4.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
python-multipart>=0.0.6
